Parsing integration for all ifupdown2 operations
Use our methods to create a standard interface selection for
ifup, ifdown, ifquery and ifreload operations.

ifquery and ifreload have a tricky implicit configuration, but I have
tried to not modify the ifupdown2 behavior in any way.

There is one changement is on the -a/--all that can be set when implicit.
Now, the command `ifreload -a -c` works for example.